Section 8: Constructor Injection
Using Spring's @Autowired annotation in order to inject the fortune service into our coach bean.

Side Note:
As of Spring Framework 4.3, an @Autowired annotation on such a constructor is no longer necessary if the target bean only defines one constructor to begin with. However, if several constructors are available, at least one must be annotated to teach the container which one to use.

Prefer to use the @Autowired annotation because it makes the code more readable.